pyproject.toml
[project]
name = "katachi"
version = "0.0.2"
description = "Python package for validating, processing and parsing directories."
authors = [{ name = "Nemanja Micovic", email = "contact.nmicovic@protonmail.com" }]
readme = "README.md"
keywords = ['python']
requires-python = ">=3.9,<4.0"
classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "adlfs>=2024.12.0",
    "fsspec>=2025.5.1",
    "loguru>=0.7.3",
    "pyyaml>=6.0.2",
    "rich>=14.0.0",
    "typer>=0.15.3",
]

[project.urls]
Homepage = "https://nmicovic.github.io/katachi/"
Repository = "https://github.com/nmicovic/katachi"
Documentation = "https://nmicovic.github.io/katachi/"

[project.scripts]
katachi = "katachi.cli:app"

[dependency-groups]
dev = [
    "pytest>=7.2.0",
    "pre-commit>=2.20.0",
    "tox-uv>=1.11.3",
    "deptry>=0.23.0",
    "mypy>=0.991",
    "pytest-cov>=4.0.0",
    "ruff>=0.11.5",
    "mkdocs>=1.4.2",
    "mkdocs-material>=8.5.10",
    "mkdocstrings[python]>=0.26.1",
    "types-pyyaml>=6.0.12.20250402",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"



[tool.mypy]
files = ["src"]
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
show_error_codes = true

[[tool.mypy.overrides]]
module = "adlfs.*"
ignore_missing_imports = true
follow_imports = "skip"
disable_error_code = ["import-untyped", "no-any-unimported", "union-attr"]

[[tool.mypy.overrides]]
module = "fsspec.*"
ignore_missing_imports = true
follow_imports = "skip"
disable_error_code = ["import-untyped", "no-any-unimported", "union-attr"]

[[tool.mypy.overrides]]
# Adding this due to ffspec not having type stubs.
module = "katachi/schema/schema_node"
disable_error_code = ["no-any-unimported"]

[[tool.mypy.overrides]]
# Adding this due to fsspec and adlfs not having type stubs.
module = "katachi/schema/importer"
disable_error_code = ["no-any-unimported", "arg-type"]

[[tool.mypy.overrides]]
# Adding this due to fsspec and adlfs not having type stubs.
module = "katachi/utils/fs_utils"
disable_error_code = ["no-any-unimported", "arg-type"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py39"
line-length = 120
fix = true

[tool.ruff.lint]
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.ruff.format]
preview = true

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["src"]

[tool.deptry.per_rule_ignores]
DEP002 = ["fsspec", "adlfs"]
